refactor(tests): Rename item-usage-journal test to journal-modal
Since this test only validates journal UI display and not actual item usage functionality, renamed it to be a standard journal modal test.

Changes:
- Renamed test file from item-usage-journal.spec.ts to journal-modal.spec.ts
- Simplified test to focus on journal modal display
- Removed unnecessary Zustand store manipulation
- Updated snapshot names to match
- Test now clearly validates journal UI, not item usage integration

Related to issue #240
