--- tests/visual/item-usage-journal.spec.ts
@@ -1,125 +1,0 @@
-import { test, expect } from '@playwright/test';
-import { seedTestData, mockApiEndpoints } from './utils/data-seeder';
-
-/**
- * Item Usage Journal Integration Visual Test
- *
- * Tests that using a significant item creates a journal entry
- * that appears in the journal modal.
- */
-
-test.describe('Item Usage Journal Integration', () => {
-  test('Should create journal entry when using significant item', async ({ page }) => {
-    // Seed all necessary test data (worlds, characters, sessions, etc.)
-    await seedTestData(page);
-    await mockApiEndpoints(page);
-
-    const sessionId = 'session-cyberpunk-ghost';
-    const worldId = 'world-cyberpunk-2077';
-    const characterId = 'char-cyberpunk-hacker';
-
-    // Navigate to the play page
-    await page.goto('/worlds/world-cyberpunk-2077/play');
-
-    // Wait for page to load
-    await page.waitForLoadState('networkidle', { timeout: 10000 });
-    await page.waitForTimeout(2000);
-
-    // Wait for active game session
-    await page.waitForSelector('[data-testid="game-session-active"]', { timeout: 10000 });
-
-    // Clear out seeded journal entries directly via Zustand store
-    await page.evaluate(() => {
-      // Access the journal store directly on window
-      const anyWindow = window as typeof window & {
-        useJournalStore?: {
-          setState: (state: any) => void;
-          getState: () => any;
-        };
-      };
-
-      if (anyWindow.useJournalStore) {
-        // Clear all entries
-        anyWindow.useJournalStore.setState({
-          entries: {},
-          currentEntryId: null,
-          error: null,
-        });
-      }
-    });
-
-    await page.waitForTimeout(500);
-
-    // Directly create a journal entry to simulate item usage via Zustand store
-    await page.evaluate(({ sessionId, worldId, characterId }) => {
-      // Access the journal store directly on window
-      const anyWindow = window as typeof window & {
-        useJournalStore?: {
-          setState: (state: any) => void;
-          getState: () => any;
-        };
-      };
-
-      if (anyWindow.useJournalStore) {
-        const journalEntry = {
-          id: 'entry-item-usage-1',
-          sessionId,
-          worldId,
-          characterId,
-          type: 'item_usage',
-          title: 'Used Crystal Key',
-          content: 'The crystal key pulses with ancient power as you hold it',
-          significance: 'minor',
-          isRead: false,
-          createdAt: new Date().toISOString(),
-          updatedAt: new Date().toISOString(),
-          relatedEntities: [{
-            type: 'item',
-            id: 'item-crystal-key',
-            name: 'Crystal Key',
-          }],
-          metadata: {
-            tags: ['item-usage', 'quest-items'],
-            automaticEntry: true,
-          },
-        };
-
-        // Get current state
-        const currentState = anyWindow.useJournalStore.getState();
-        const entries = { ...currentState.entries };
-
-        // Add entry for this session
-        if (!entries[sessionId]) {
-          entries[sessionId] = [];
-        }
-        entries[sessionId].push(journalEntry);
-
-        // Update store
-        anyWindow.useJournalStore.setState({ entries });
-      }
-    }, { sessionId, worldId, characterId });
-
-    await page.waitForTimeout(1000);
-
-    // Look for journal floating button
-    const journalButton = page.getByRole('button', { name: /open journal/i });
-    await expect(journalButton).toBeVisible({ timeout: 10000 });
-
-    // Click the journal button to open the modal
-    await journalButton.click();
-
-    // Wait for journal modal to open
-    await page.waitForSelector('[role="dialog"]', { timeout: 5000 });
-
-    // Verify journal modal is visible
-    const journalModal = page.locator('[role="dialog"]');
-    await expect(journalModal).toBeVisible();
-
-    // Take screenshot of journal with item usage entry
-    await page.waitForTimeout(500); // Let modal animation complete
-    await expect(page).toHaveScreenshot('item-usage-journal.png', {
-      fullPage: false,
-      threshold: 0.3,
-    });
-  });
-});
