+++ tests/visual/journal-modal.spec.ts-snapshots/journal-modal-chromium-darwin.png
